check token valid
